# Conflicts:
#	app/config/services.yml
#	src/RocketSeller/TwoPickBundle/Mailer/TwigSwiftMailer.php
#	src/RocketSeller/TwoPickBundle/Resources/views/BackOffice/backOfficeBase.html.twig

+++ src/RocketSeller/TwoPickBundle/Controller/HighTechRestController.php
@@ -98,7 +98,7 @@ class HighTechRestController extends FOSRestController
         $mandatory = array();
 
         // Set all the parameters info.
-        $regex['numeroRadicado'] = '([0-9])+';
+        $regex['numeroRadicado'] = '(.)+';
         $mandatory['numeroRadicado'] = true;
         $regex['estado'] = '([0-9])+';
         $mandatory['estado'] = true;
@@ -180,6 +180,21 @@ class HighTechRestController extends FOSRestController
 
             $this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($context);
 
+            //push notification
+            $message = "¡Hemos debitado tu cuenta con éxito!";
+            $title = "Symplifica";
+            $longMessage = "¡Hemos debitado tu cuenta con éxito! Te avisaremos en cuanto se hayan realizado los pagos correspondientes";
+
+            $request = new Request();
+            $request->setMethod("POST");
+            $request->request->add(array(
+                "idUser" => $rejectedPurchaseOrderDescription->getPurchaseOrders()->getIdUser(),
+                "title" => $title,
+                "message" => $message,
+                "longMessage" => $longMessage
+            ));
+            $pushNotificationService = $this->get('app.symplifica_push_notification');
+            $result = $pushNotificationService->postPushNotificationAction($request);
         } else {
             //nicetohave buscar este ID
             $paymethodId = $dis->getPayMethodId();
@@ -200,7 +215,23 @@ class HighTechRestController extends FOSRestController
                 'phone'=>$dis->getIdUser()->getPersonPerson()->getPhones()->first()->getPhoneNumber(),
 	              'value'=>$dis->getValue()
             );
-            $this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($contextBack);
+            //$this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($contextBack);
+
+            //push notification
+            $message = "Hubo un inconveniente al debitar tu cuenta";
+            $title = "Symplifica";
+            $longMessage = "Tuvimos un inconveniente al realizar el débito de tu cuenta, nos pondremos en contacto.";
+
+            $request = new Request();
+            $request->setMethod("POST");
+            $request->request->add(array(
+                "idUser" => $rejectedPurchaseOrderDescription->getPurchaseOrders()->getIdUser(),
+                "title" => $title,
+                "message" => $message,
+                "longMessage" => $longMessage
+            ));
+            $pushNotificationService = $this->get('app.symplifica_push_notification');
+            $result = $pushNotificationService->postPushNotificationAction($request);
 
             $pos = $this->getDoctrine()->getRepository("RocketSellerTwoPickBundle:PurchaseOrdersStatus")->findOneBy(array('idNovoPay' => 'P1'));
             //$realtoPay->setPurchaseOrdersStatus($procesingStatus);
@@ -304,6 +335,22 @@ class HighTechRestController extends FOSRestController
                 $context['documentName']='Comprobante '.date_format(new DateTime(),'d-m-y H:i:s').'.pdf';
                 $this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($context);
 
+                //push notification
+                $message = "¡Hemos realizado tu pago con éxito!";
+                $title = "Symplifica";
+                $longMessage = "¡Hemos realizado tu pago con éxito! Puedes entrar y obtener sus comprobantes";
+
+                $request = new Request();
+                $request->setMethod("POST");
+                $request->request->add(array(
+                    "idUser" => $rejectedPurchaseOrderDescription->getPurchaseOrders()->getIdUser(),
+                    "title" => $title,
+                    "message" => $message,
+                    "longMessage" => $longMessage
+                ));
+                $pushNotificationService = $this->get('app.symplifica_push_notification');
+                $result = $pushNotificationService->postPushNotificationAction($request);
+
                 //here i create the comprobante
                 /* This is not longer necesary
 
@@ -384,6 +431,22 @@ class HighTechRestController extends FOSRestController
 				        $context['documentName']= 'Comprobante pago aportes '.date_format(new DateTime(),'d-m-y H:i:s').'.pdf';
 				        $this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($context);
 
+                        //push notification
+                        //¡El pago de los aportes a seguridad social se ha realizado con éxito!
+                        $message = "¡Hemos realizado tu pago con éxito!";
+                        $title = "Symplifica";
+                        $longMessage = "¡Hemos realizado El pago de los aportes a seguridad social con éxito!";
+
+                        $request = new Request();
+                        $request->setMethod("POST");
+                        $request->request->add(array(
+                            "idUser" => $rejectedPurchaseOrderDescription->getPurchaseOrders()->getIdUser(),
+                            "title" => $title,
+                            "message" => $message,
+                            "longMessage" => $longMessage
+                        ));
+                        $pushNotificationService = $this->get('app.symplifica_push_notification');
+                        $result = $pushNotificationService->postPushNotificationAction($request);
 			        }
 		        }
 
@@ -419,7 +482,7 @@ class HighTechRestController extends FOSRestController
                 'idPOD'=>$rejectedPurchaseOrderDescription->getIdPurchaseOrdersDescription(),
                 'value'=>$value
             );
-            $this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($contextBack);
+            //$this->get('symplifica.mailer.twig_swift')->sendEmailByTypeMessage($contextBack);
 
             $notification= new Notification();
             $notification->setAccion("Ver");
@@ -435,6 +498,22 @@ class HighTechRestController extends FOSRestController
             $em->persist($notification);
             $em->flush();
 
+            //push notification
+            $message = "Hubo un inconveniente al realizar tu pago";
+            $title = "Symplifica";
+            $longMessage = "Tuvimos un inconveniente al realizar tu pago, puedes realizar el cambio de tu método de pago";
+
+            $request = new Request();
+            $request->setMethod("POST");
+            $request->request->add(array(
+                "idUser" => $rejectedPurchaseOrderDescription->getPurchaseOrders()->getIdUser(),
+                "title" => $title,
+                "message" => $message,
+                "longMessage" => $longMessage,
+                "page" => 'TransactionsInProcessPage'
+            ));
+            $pushNotificationService = $this->get('app.symplifica_push_notification');
+            $result = $pushNotificationService->postPushNotificationAction($request);
         }
         $em->persist($pay);
         $em->flush();
@@ -579,8 +658,8 @@ class HighTechRestController extends FOSRestController
 			$request->setMethod("PUT");
 			$request->request->add(array(
 				"radicatedNumber"=> $parameters['radicatedNumber'],
-				"registerState"=> isset($parameters['registerState']) && $parameters['registerState'] != NULL ? $parameters['registerState'] : "",
-				"errorLog" => isset($oarameters['errorLog']) &&  $oarameters['errorLog'] != NULL ? $parameters['errorLog'] : "",
+				"registerState"=> isset($parameters['registerState']) && $parameters['registerState'] != NULL ? $parameters['registerState'] : -1,
+				"errorLog" => isset($parameters['errorLog']) &&  $parameters['errorLog'] != NULL ? $parameters['errorLog'] : "",
 				"errorMessage" => isset($parameters['errorMessage']) && $parameters['errorMessage'] ? $parameters['errorMessage'] : ""
 			));
 
@@ -627,11 +706,17 @@ class HighTechRestController extends FOSRestController
 		$mandatory['payrollNumber'] = false;
 
 		$this->validateParamters($parameters, $regex, $mandatory);
-
+		
+		/*There are 4 cases...
+		The file uploaded (payrollState = 0) and was no problems, errorLog == null
+		The file uploaded (payrollState = 0) and was warnings or errors, errorLog != null
+		The file was not uploaded (payrollState != 0) and was a big problem, errorLog = null
+		The file was not uploaded (payrollState != 0) and was a specific problem , errorLog != null*/
+		
 		$request->setMethod("PUT");
 		$request->request->add(array(
 			"radicatedNumber"=> $parameters['radicatedNumber'],
-			"planillaState"=> isset($parameters['payrollState']) && $parameters['payrollState'] != NULL ? $parameters['payrollState'] : "",
+			"planillaState"=> isset($parameters['payrollState']) && $parameters['payrollState'] != NULL ? $parameters['payrollState'] : -1,
 			"errorLog" => isset($parameters['errorLog']) && $parameters['errorLog'] != NULL ? $parameters['errorLog'] : "",
 			"planillaNumber" => isset($parameters['payrollNumber']) && $parameters['payrollNumber'] != NULL ? $parameters['payrollNumber'] : "",
 			"errorMessage" => isset($parameters['errorMessage']) && $parameters['errorMessage'] != NULL ? $parameters['errorMessage'] : ""
@@ -683,7 +768,7 @@ class HighTechRestController extends FOSRestController
 		}
 
 		//This means the user was created succesfully
-		if( $parameters['registerState'] == 0 && $parameters['errorLog'] == "" && $parameters['errorMessage'] == "" ){
+		if( $parameters['registerState'] == 0 ){
 			$purchaseOrdersStatus = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersStatus')->findOneBy(array('idNovoPay' => 'InsPil-InsOk'));
 			$singleTransaction->setPurchaseOrdersStatus($purchaseOrdersStatus);
 
@@ -694,49 +779,64 @@ class HighTechRestController extends FOSRestController
 			$em->flush();
 
 		}
-		else if( $parameters['registerState'] != NULL && $parameters['errorLog'] != "" && $parameters['errorMessage'] != "" ){
+		else if( $parameters['registerState'] != -1 ){
 			$purchaseOrdersStatus = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersStatus')->findOneBy(array('idNovoPay' => 'InsPil-InsRec'));
 			$documentType = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:DocumentType')->findOneBy(array('docCode' => 'EOCP'));
 			$singleTransaction->setPurchaseOrdersStatus($purchaseOrdersStatus);
 
 			$employer = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:Employer')->findOneBy(array('existentPila' => $singleTransaction->getIdTransaction()));
 
-			/** @var TransactionState $transactionState */
-			$transactionState = new TransactionState();
-			/** @var Document $document */
-			$document = new Document();
-			$document->setName("Enlace operativo employer creation error " . $employer->getIdEmployer());
-			$document->setStatus(1);
-			$document->setDocumentTypeDocumentType($documentType);
-
-			$filename = "tempErrorImg.zip";
-			$file = "uploads/temp/$filename";
-
-			file_put_contents($file, base64_decode($parameters['errorLog']));
-
-			$mediaManager = $this->container->get('sonata.media.manager.media');
-      $media = $mediaManager->create();
-      $media->setBinaryContent($file);
-      $media->setProviderName('sonata.media.provider.file');
-      $media->setName($document->getName());
-      $media->setProviderStatus(Media::STATUS_OK);
-      $media->setContext('person');
-      $media->setDocumentDocument($document);
-
-			$document->setMediaMedia($media);
-
-			$em->persist($document);
-
-			$transactionState->setDocument($document);
-			$transactionState->setLog($parameters['errorMessage']);
+			if($singleTransaction->getTransactionState() != NULL){
+				$transactionState = $singleTransaction->getTransactionState();
+			}
+			else{
+				/** @var TransactionState $transactionState */
+				$transactionState = new TransactionState();
+			}
+			
+			if($parameters['errorLog'] != "" ){
+				/** @var Document $document */
+				$document = new Document();
+				$document->setName("Enlace operativo employer creation error " . $employer->getIdEmployer());
+				$document->setStatus(1);
+				$document->setDocumentTypeDocumentType($documentType);
+				
+				$filename = "tempErrorImg.zip";
+				$file = "uploads/temp/$filename";
+				
+				file_put_contents($file, base64_decode($parameters['errorLog']));
+				
+				$mediaManager = $this->container->get('sonata.media.manager.media');
+				$media = $mediaManager->create();
+				$media->setBinaryContent($file);
+				$media->setProviderName('sonata.media.provider.file');
+				$media->setName($document->getName());
+				$media->setProviderStatus(Media::STATUS_OK);
+				$media->setContext('person');
+				$media->setDocumentDocument($document);
+				
+				$document->setMediaMedia($media);
+				
+				$em->persist($document);
+				
+				$transactionState->setDocument($document);
+			}
+			
+			if($parameters['errorMessage'] != ""){
+				$transactionState->setLog($parameters['errorMessage']);
+			}
+			
 			$transactionState->setOriginTransaction($singleTransaction);
 			$em->persist($transactionState);
 
 			$singleTransaction->setTransactionState($transactionState);
 			$em->persist($singleTransaction);
 			$em->flush();
-
-			unlink($file);
+			
+			if($parameters['errorLog'] != "" ) {
+				unlink($file);
+			}
+			
 		}
 		// Succesfull operation.
 		$view = View::create();
@@ -787,9 +887,12 @@ class HighTechRestController extends FOSRestController
 			$view->setData(array("returnCode" => 404 , "returnDescription" => "Radicated number not found"));
 			return $view;
 		}
-
-		//This means the planilla was created succesfully
-		if( $parameters['planillaState'] == 0 && $parameters['errorLog'] == "" && $parameters['errorMessage'] == "" ){
+		
+		$estadoPlanilla = $parameters['planillaState'];
+		$errorLog = $parameters['errorLog'];
+		
+		if( $estadoPlanilla == 0 && $errorLog == "" ){
+			//This means the planilla was created succesfully
 			$purchaseOrdersStatus = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersStatus')->findOneBy(array('idNovoPay' => 'CarPla-PlaOK'));
 			$singleTransaction->setPurchaseOrdersStatus($purchaseOrdersStatus);
 
@@ -801,26 +904,34 @@ class HighTechRestController extends FOSRestController
 			$em->flush();
 
 		}
-		else if( $parameters['planillaState'] != NULL && $parameters['errorLog'] != "" && $parameters['errorMessage'] != "" ){
-			$purchaseOrdersStatus = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersStatus')->findOneBy(array('idNovoPay' => 'CarPla-PlaErr'));
+		elseif ( $estadoPlanilla == 0 && $errorLog != ""){
+			//This means the planilla was created succesfully but has warnings
+			$purchaseOrdersStatus = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersStatus')->findOneBy(array('idNovoPay' => 'CarPla-PlaWar'));
 			$documentType = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:DocumentType')->findOneBy(array('docCode' => 'EOIE'));
 			$singleTransaction->setPurchaseOrdersStatus($purchaseOrdersStatus);
-
+			
 			$purchaseOrderDescription = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersDescription')->findOneBy(array('uploadedFile' => $singleTransaction->getIdTransaction()));
-
-			/** @var TransactionState $transactionState */
-			$transactionState = new TransactionState();
+			$purchaseOrderDescription->setEnlaceOperativoFileName($parameters['planillaNumber']);
+			
+			if($singleTransaction->getTransactionState() != NULL){
+				$transactionState = $singleTransaction->getTransactionState();
+			}
+			else{
+				/** @var TransactionState $transactionState */
+				$transactionState = new TransactionState();
+			}
+			
 			/** @var Document $document */
 			$document = new Document();
-			$document->setName("Enlace operativo upload file error " . $purchaseOrderDescription->getIdPurchaseOrdersDescription());
+			$document->setName("Enlace operativo uploaded file warning " . $purchaseOrderDescription->getIdPurchaseOrdersDescription());
 			$document->setStatus(1);
 			$document->setDocumentTypeDocumentType($documentType);
-
-			$filename = "tempErrorImg.zip";
+			
+			$filename = "tempWarningImg.zip";
 			$file = "uploads/temp/$filename";
-
+			
 			file_put_contents($file, base64_decode($parameters['errorLog']));
-
+			
 			$mediaManager = $this->container->get('sonata.media.manager.media');
 			$media = $mediaManager->create();
 			$media->setBinaryContent($file);
@@ -829,21 +940,81 @@ class HighTechRestController extends FOSRestController
 			$media->setProviderStatus(Media::STATUS_OK);
 			$media->setContext('person');
 			$media->setDocumentDocument($document);
-
+			
 			$document->setMediaMedia($media);
-
+			
 			$em->persist($document);
-
+			
 			$transactionState->setDocument($document);
 			$transactionState->setLog($parameters['errorMessage']);
+			$transactionState->setOriginTransaction($singleTransaction);
+			$em->persist($transactionState);
+			
+			$singleTransaction->setTransactionState($transactionState);
+			$em->persist($singleTransaction);
+			
+			$em->persist($purchaseOrderDescription);
+			$em->flush();
+			
+			unlink($file);
+		}
+		else if( $parameters['planillaState'] != -1  ){
+			$purchaseOrdersStatus = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersStatus')->findOneBy(array('idNovoPay' => 'CarPla-PlaErr'));
+			$documentType = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:DocumentType')->findOneBy(array('docCode' => 'EOIE'));
+			$singleTransaction->setPurchaseOrdersStatus($purchaseOrdersStatus);
+			
+			$purchaseOrderDescription = $this->getDoctrine()->getRepository('RocketSellerTwoPickBundle:PurchaseOrdersDescription')->findOneBy(array('uploadedFile' => $singleTransaction->getIdTransaction()));
+			
+			if($singleTransaction->getTransactionState() != NULL){
+				$transactionState = $singleTransaction->getTransactionState();
+			}
+			else{
+				/** @var TransactionState $transactionState */
+				$transactionState = new TransactionState();
+			}
+			
+			if($parameters['errorLog'] != "" ) {
+				/** @var Document $document */
+				$document = new Document();
+				$document->setName("Enlace operativo upload file error " . $purchaseOrderDescription->getIdPurchaseOrdersDescription());
+				$document->setStatus(1);
+				$document->setDocumentTypeDocumentType($documentType);
+				
+				$filename = "tempErrorImg.zip";
+				$file = "uploads/temp/$filename";
+				
+				file_put_contents($file, base64_decode($parameters['errorLog']));
+				
+				$mediaManager = $this->container->get('sonata.media.manager.media');
+				$media = $mediaManager->create();
+				$media->setBinaryContent($file);
+				$media->setProviderName('sonata.media.provider.file');
+				$media->setName($document->getName());
+				$media->setProviderStatus(Media::STATUS_OK);
+				$media->setContext('person');
+				$media->setDocumentDocument($document);
+				
+				$document->setMediaMedia($media);
+				
+				$em->persist($document);
+				
+				$transactionState->setDocument($document);
+			}
+			
+			if($parameters['errorMessage'] != ""){
+				$transactionState->setLog($parameters['errorMessage']);
+			}
+			
 			$transactionState->setOriginTransaction($singleTransaction);
 			$em->persist($transactionState);
 
 			$singleTransaction->setTransactionState($transactionState);
 			$em->persist($singleTransaction);
 			$em->flush();
-
-			unlink($file);
+			
+			if($parameters['errorLog'] != "" ) {
+				unlink($file);
+			}
 		}
 		// Succesfull operation.
 		$view = View::create();
